Test/Xsrc/requirements.txt

(venv) C:\Users\bzelb\PycharmProjects\experimentalHands>pip list
Package         Version
--------------- ---------
absl-py         0.12.0
attrs           20.3.0
beautifulsoup4  4.9.3
certifi         2020.12.5
chardet         4.0.0
cycler          0.10.0
dataclasses     0.6
docopt          0.6.2
idna            2.10
Js2Py           0.71
kiwisolver      1.3.1
matplotlib      3.4.2
mediapipe       0.8.3.1
numpy           1.20.2
opencv-python   4.5.1.48
packaging       20.9
Pillow          8.2.0
pip             21.1.1
pipwin          0.5.1
protobuf        3.15.8
PyAudio         0.2.11
pyjsparser      2.7.1
pyparsing       2.4.7
PyPrind         2.11.3
pySmartDL       1.3.4
python-dateutil 2.8.1
pytz            2021.1
requests        2.25.1
setuptools      56.0.0
six             1.15.0
soupsieve       2.2.1
tzlocal         2.1
urllib3         1.26.4
wheel           0.36.2
python-rtmidi    ~^
WARNING: You are using pip version 21.1.1; however, version 21.1.2 is available.
